+++ scripts/pack/logtalk/pack.pl
@@ -1,0 +1,8 @@
+name(logtalk).
+title('Logtalk - Object-Oriented Logic Programming Language').
+version('3.00.0').
+author('Paulo Moura', 'pmoura@logtalk.org').
+packager('Paulo Moura', 'pmoura@logtalk.org').
+maintainer('Paulo Moura', 'pmoura@logtalk.org').
+home('http://logtalk.org/').
+download('http://logtalk.org/files/swi-prolog/packs/logtalk-3.00.0.tgz').