+++ scripts/pack/logtalk/prolog/logtalk.pl
@@ -1,0 +1,34 @@
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  This file is part of Logtalk <http://logtalk.org/>  
+%  Copyright (c) 1998-2015 Paulo Moura <pmoura@logtalk.org>
+%
+%  Support for load-on-demand using SWI Prolog 6.6.0 and later versions
+%  Last updated on January 15, 2015
+%
+%  This program is free software: you can redistribute it and/or modify
+%  it under the terms of the GNU General Public License as published by
+%  the Free Software Foundation, either version 3 of the License, or
+%  (at your option) any later version.
+%  
+%  This program is distributed in the hope that it will be useful,
+%  but WITHOUT ANY WARRANTY; without even the implied warranty of
+%  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+%  GNU General Public License for more details.
+%  
+%  You should have received a copy of the GNU General Public License
+%  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+%  
+%  Additional licensing terms apply per Section 7 of the GNU General
+%  Public License 3. Consult the `LICENSE.txt` file for details.
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+:- module(logtalk, []).
+
+:-	prolog_load_context(directory, Directory),
+	atom_concat(Directory, '/../logtalk-3.00.0-stable/', Location),
+	setenv('LOGTALKHOME', Location),
+	setenv('LOGTALKUSER', Location),
+	user:load_files('../logtalk-3.00.0-stable/integration/logtalk_swi.pl').